requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
colorama==0.4.6
distro==1.9.0
Django==4.2.16
django-environ==0.13.0
djangorestframework==3.17.1
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
numpy==2.2.6
openai==2.26.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.0
pandas==2.2.3
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyodbc==5.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
selenium==4.41.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
tqdm==4.67.1
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
Pillow==12.1.1
